Finalizado juego de coche y ajustes de dificultad

+++ CarGame/Assets/Scripts/ConfigurationSave/ConfigurationData.cs
@@ -7,5 +7,5 @@ public class ConfigurationData
 {
     public int Laps;
     public int AnguloDeJuego;
-    public int Dificultad;
+    public float Dificultad;
 }
